Userspace server and pass in ul instead of int

+++ dpdk/server_helpers.cpp
@@ -50,8 +50,8 @@ std::string compute_response(struct request *req) {
     std::string resp;
 
     if (req->type == REQUEST_TYPE_GET) {
-        char *value = server_grab_value(stoi(req->key), 1, ROOT_NODE_OFFSET);
-        
+        char *value = server_grab_value(std::stoul(req->key), 1, ROOT_NODE_OFFSET);
+
         if (value != nullptr)
             resp = std::string("VALUE ") + std::string(value) + std::string(" ") + req->id;
         else
